finished TODOs
