+++ src/actions/abilities.ts
@@ -4,12 +4,12 @@ export const setFocusedAbility = (focusedAbility: string | null) => ({
   payload: focusedAbility
 })
 
-export const SET_ABILITY_COMPLETE_ON_CURRENT_TURN = 'SET_ABILITY_COMPLETE_ON_CURRENT_TURN'
-// TODO: Implement Reducer (Renamed. See ABILITY_COMPLETED)
-export const setAbilityCompleteOnCurrentTurn = (abilityName: string | null) => ({
-  type: SET_ABILITY_COMPLETE_ON_CURRENT_TURN,
-  payload: abilityName
-})
+// export const SET_ABILITY_COMPLETE_ON_CURRENT_TURN = 'SET_ABILITY_COMPLETE_ON_CURRENT_TURN'
+// // TODO: Implement Reducer (Renamed. See ABILITY_COMPLETED)
+// export const setAbilityCompleteOnCurrentTurn = (abilityName: string | null) => ({
+//   type: SET_ABILITY_COMPLETE_ON_CURRENT_TURN,
+//   payload: abilityName
+// })
 
 export const SET_DEFAULT_ABILITY = 'SET_DEFAULT_ABILITY' // TODO: implement reducer (Done)
 export const setDefaultAbility = (defaultAbilityName: string | null) => ({
